backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
langchain-anthropic==0.3.3
langgraph==0.2.60
langchain-core==0.3.25
supabase==2.10.0
httpx==0.28.1
python-dotenv==1.0.1
